+++ examples/slidebook/package.json
@@ -8,8 +8,9 @@
     "start": "slidebook start"
   },
   "dependencies": {
-    "@slidebook/core": "file:..\\..\\packages\\core\\slidebook-core-0.1.23.tgz",
-    "@slidebook/image": "file:..\\..\\packages\\image\\slidebook-image-0.1.78.tgz",
+    "@slidebook/core": "file:..\\..\\packages\\core\\slidebook-core-0.1.24.tgz",
+    "@slidebook/image": "file:..\\..\\packages\\image\\slidebook-image-0.1.84.tgz",
+    "@slidebook/server": "file:..\\..\\packages\\server\\slidebook-server-0.1.8.tgz",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
     "framer-motion": "12.6.5",
@@ -19,7 +20,6 @@
     "tailwind-merge": "3.0.2",
     "cross-env": "7.0.3",
     "dotenv": "16.4.7",
-    "socket.io": "4.8.1",
     "tw-animate-css": "1.2.5"
   },
   "devDependencies": {
