+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@slidebook/core",
-  "version": "0.1.23",
+  "version": "0.1.24",
   "private": true,
   "files": [
     "lib"
