--- packages/core/src/lib/authenticate.ts
@@ -1,3 +1,0 @@
-export const isAuthenticated = (password: string) => {
-    return password === process.env.PASSWORD;
-};
