+++ packages/server/lib/authenticate.js
@@ -1,0 +1,3 @@
+export const isAuthenticated = (password) => {
+    return password === process.env.PASSWORD;
+};
